proxy: support SOCK_STREAM sockets
macOS does not support SOCK_SEQPACKET so we have to make use of SOCK_STREAM.
That means we have to take care of framing. Thankfully, the SCM_RIGHTS
buffer will always be complete and ordered so we send that with a header
indicating the number of bytes for the packet followed by the packet.

Co-authored-by: Lucas Amaral <lucaaamaral@gmail.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1600>